refactor(ai-diagnostic): extract IsolationForest from anomaly_detection
Moved the self-contained Isolation Forest / Isolation Tree scorer (Liu et al.,
pure numpy) into service/isolation_forest.py, separate from the metric-ingestion
AnomalyDetector orchestration. anomaly_detection.py re-exports both classes, so
importers are unchanged. Decomposed under the existing green suite — the 16
anomaly_detection tests (which import IsolationForest directly) stay green.
Randomness still flows through the shared seeded RNG for reproducibility.
Dropped now-dead imports (random, get_rng/_rng). 718 -> 565 LOC.
